:truck: Changed structure of Backend
